Move FollowSet to it's own class.
